requirements.txt
Pillow
numpy
pdf2image
p_tqdm
